Ya pinta el HexGrid. Pero, solo algunas veces :C

+++ uad_progra3/Include/CHexGridCell.h
@@ -7,7 +7,7 @@
 class CHexGridCell
 {
 public:
-	CHexGridCell(int x, int y);
+	CHexGridCell(float x, float y);
 	~CHexGridCell();
 
 	CVector3 getCenter();
